Create ThreadedBinaryTree.c
This code is the implementation of a subtype of binary trees which is threaded binary tree in C language. As the name says this type has left and right threads for each node that connects their inorder predecessor  and inorder successor respectively which makes it easier to implement inorder traversal across the tree.